requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==1.5.3
scikit-learn==1.0.2
numpy==1.21.6
joblib==1.2.0
torch==2.8.0
torchvision==0.23.0
transformers==4.56.2
Pillow==10.4.0